requirements.txt
asgiref==3.4.1
boto3==1.19.7
botocore==1.22.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
decorator==5.1.0
Django==3.2.8
django-rest-knox==4.1.0
djangorestframework==3.12.4
dynaconf==3.1.7
hvac==0.11.2
idna==3.3
Jinja2==3.0.2
jmespath==0.10.0
knox==0.1.14
loguru==0.5.3
MarkupSafe==2.0.1
pycparser==2.20
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
validators==0.18.2
win32-setctime==1.0.3
